+++ Informatik/Informatik/util/SDL_Util.cpp
@@ -65,8 +65,8 @@ float drawText(const char *text, uint32_t color, SDL_Rect dst, cachedTexture &te
     {
         int neww = (int) ((float) texture_cache.width * texture_cache.scale);
         int newh = (int) ((float) texture_cache.height * texture_cache.scale);
-        dst.x += (dst.w - neww) / 2.0f;
-        dst.y += (dst.h - newh) / 2.0f;
+        dst.x += (dst.w - neww) / 2;
+        dst.y += (dst.h - newh) / 2;
     }
     dst.w = neww;
     dst.h = newh;
@@ -225,11 +225,11 @@ void brighten(SDL_Surface *surface, float multiplier)
     for(int i = 0; i < surface->w * surface->h; i++)
     {
         uint32_t cp = ((uint32_t*) surface->pixels)[i];
-        uint32_t r = (cp & fml->Rmask) * (multiplier + 1);
+        uint32_t r = (uint32_t) ((cp & fml->Rmask) * (multiplier + 1.0f));
         if (r > fml->Rmask) r = fml->Rmask;
-        uint32_t g = (cp & fml->Gmask) * (multiplier + 1);
+        uint32_t g = (uint32_t) ((cp & fml->Gmask) * (multiplier + 1.0f));
         if (g > fml->Gmask) g = fml->Gmask;
-        uint32_t b = (cp & fml->Bmask) * (multiplier + 1);
+        uint32_t b = (uint32_t) ((cp & fml->Bmask) * (multiplier + 1.0f));
         if (b > fml->Bmask) b = fml->Bmask;
         pixels[i] = (cp & fml->Amask) | r | g | b;
     }
